refactor systemTools
