+++ packages/rnv/src/cli/tools.js
@@ -20,7 +20,7 @@ import {
     cleanPlatformIfRequired,
 } from '../common';
 import { IOS } from '../constants';
-import { executeAsync, execCLI } from '../exec';
+import { executeAsync, execCLI } from '../systemTools/exec';
 import { executePipe } from '../projectTools/buildHooks';
 import {
     packageAndroid,
